+++ app/schemas/stock.py
@@ -1,6 +1,5 @@
 import uuid
 from datetime import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -9,7 +8,7 @@ class StockRead(BaseModel):
     id: uuid.UUID
     product_id: uuid.UUID
     location_id: uuid.UUID
-    quantity: Decimal
+    quantity: float
     updated_at: datetime
 
     model_config = {"from_attributes": True}
@@ -18,4 +17,4 @@ class StockRead(BaseModel):
 class StockAdjust(BaseModel):
     product_id: uuid.UUID
     location_id: uuid.UUID
-    quantity: Decimal
+    quantity: float
